genai-worker/requirements.txt
pika==1.3.2
requests==2.32.3
google-genai==1.0.0
python-dotenv==1.0.1
loguru==0.7.2